refactor: 주문 테이블의 빈 자리 여부 변경 시 주문을 주입받아 체크하는 validator 의 역할을 주문에서 처리하도록 수정하여 의존성 제거 추가
